+++ app/src/main/java/com/agileburo/anytype/ui/auth/account/SetupSelectedAccountFragment.kt
@@ -3,6 +3,7 @@ package com.agileburo.anytype.ui.auth.account
 import android.os.Bundle
 import android.view.View
 import android.view.animation.AnimationUtils
+import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import com.agileburo.anytype.R
 import com.agileburo.anytype.di.common.componentManager
@@ -10,13 +11,24 @@ import com.agileburo.anytype.presentation.auth.account.SetupSelectedAccountViewM
 import com.agileburo.anytype.presentation.auth.account.SetupSelectedAccountViewModelFactory
 import com.agileburo.anytype.ui.auth.Keys
 import com.agileburo.anytype.ui.base.NavigationFragment
-import kotlinx.android.synthetic.main.fragment_setup_new_account.*
+import kotlinx.android.synthetic.main.fragment_setup_new_account.icon
+import kotlinx.android.synthetic.main.fragment_setup_selected_account.*
 import javax.inject.Inject
 
-class SetupSelectedAccountFragment : NavigationFragment(R.layout.fragment_setup_selected_account) {
+open class SetupSelectedAccountFragment :
+    NavigationFragment(R.layout.fragment_setup_selected_account) {
 
     @Inject
     lateinit var factory: SetupSelectedAccountViewModelFactory
+
+    private val animation by lazy {
+        AnimationUtils.loadAnimation(requireContext(), R.anim.rotation)
+    }
+
+    private val errorObserver = Observer<String> {
+        error.text = it
+        animation.cancel()
+    }
 
     private val vm by lazy {
         ViewModelProviders
@@ -31,16 +43,17 @@ class SetupSelectedAccountFragment : NavigationFragment(R.layout.fragment_setup_
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        icon.startAnimation(AnimationUtils.loadAnimation(requireContext(), R.anim.rotation))
+        icon.startAnimation(animation)
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        setupNavigation()
+        subscribe()
     }
 
-    private fun setupNavigation() {
-        vm.observeNavigation().observe(this, navObserver)
+    private fun subscribe() {
+        vm.observeNavigation().observe(viewLifecycleOwner, navObserver)
+        vm.error.observe(viewLifecycleOwner, errorObserver)
     }
 
     override fun injectDependencies() {